Fixed handling of multi-line descriptions #2973
