+++ src/PSRule.Rules.Azure/Data/Policy/ILazyValue.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Rules.Azure.Data.Policy
+{
+    internal interface ILazyValue
+    {
+        object GetValue(PolicyAssignmentVisitor.PolicyAssignmentContext context);
+    }
+}
